removed virtual tables from gfx objects, struct size reductions

+++ WickedEngine/wiGraphicsDevice_Vulkan.cpp
@@ -3999,7 +3999,7 @@ using namespace vulkan_internal;
 				internal_state->resource,
 				&memory_requirements
 			);
-			buffer->sparse_page_size = memory_requirements.alignment;
+			buffer->sparse_page_size = (uint32_t)memory_requirements.alignment;
 		}
 		else
 		{
@@ -4341,7 +4341,7 @@ using namespace vulkan_internal;
 				internal_state->resource,
 				&memory_requirements
 			);
-			texture->sparse_page_size = memory_requirements.alignment;
+			texture->sparse_page_size = (uint32_t)memory_requirements.alignment;
 
 			uint32_t sparse_requirement_count = 0;
